Moved the filterdata to utils folder.
