+++ src/libs/utils/FilterData.ts
@@ -1,0 +1,29 @@
+const filterData = {
+    productCategory: [
+      'All Products',
+      'Antioxidants',
+      'Broad Spectrum SPF',
+      'Cleansers & Toners',
+      'Eye, Neck, Lip',
+      'Masks',
+      'Moisturizers',
+      'Retinols',
+      'Serums',
+      'Solution Sets',
+      'Peels',
+      'Peel Alternative',
+      'Retinoid Treatments',
+      'Therapeutic Masks',
+      'Treatment Enhancements',
+      'Backbar Sizes',
+      'Backbar Tools',
+      'Body Treatments',
+      'Exfoliants',
+    ],
+    bestSeller: ['Best Seller'],
+    productTypes: ['Backbar', 'Retail', 'Sample'],
+    skinConcern: ['Brightening', 'Acne', 'Aging', 'Discolorations', 'Preventative', 'Sensitive Skin'],
+    skinType: ['Combination', 'Dry', 'Normal', 'Oily', 'Sensitive'],
+  };
+
+  export default filterData;
